Move non_differentiable_arg_names from autograd functions to differentiability_info.
There was a similar problem with as with output_differentiability: if an autograd function is not generated, we just throw away this information.
This doesn't particularly matter in this case, because no codegen depends on it, I'm just moving all argument differentiability arguments off of the autograd function.

gh-metadata: pytorch pytorch 19425 gh/gchanan/18/head